Create `FeedStoreSpecs` that must be implemented by any `FeedStore` test to guarantee it meets specs
